Counter to track files, store overall result in db
It is difficult to ensure that all files are being processed
without a counter to track them.
There is a need to store the overall result "PASSED" or "FAILED"
into the database, so that this information can be displayed on
the dashboard.

The files are being tracked with variables that count the number
of files dropped and processed, which get printed so that the user
is aware of the progress.
The information is added to the database by invoking functions
in db.py

Signed-off-by: Angelina Vu <vuangelina72@gmail.com>