abipkgdiff kernel: kernel packages need debuginfo
In order to run abipkgdiff with a kernel package, that package
must be accompanied by a debuginfo package.

"kernel" was removed from the list of filter words, since it was
found that abipkgdiff works fine with kernel packages, as long as
there is a corresponding debuginfo package.
abicheck.generate_package_json drops any kernel packages
that don't have a debuginfo package.

Signed-off-by: Angelina Vu <vuangelina72@gmail.com>