RPM Download Retry Logic: Fix bug with empty urls
Implement retry logic into url query and download. These steps are
retried up to 3 times if the url is an empty string or if the download
fails.
If after 3 attempts, the download still fails, the program moves on to
the next package.

This is to fix the issue of the url query sometimes returning an empty
url, causing the program to stop execution.

Signed-off-by: Angelina Vu <vuangelina72@gmail.com>